feat(spec-063 D4): one INSERT BULK builder — and three tests that had gone stale
CTAS emitted only `WITH (TABLOCK)` and never `ROWS_PER_BATCH`, so the server was
told the batch size for a COPY and left to guess it for a CTAS. Nobody chose
that; it is what two builders do. There is one now, and both sides produce
byte-identical text — verified live:

    INSERT BULK [dbo].[D4Ctas] ([id] bigint) WITH (TABLOCK, ROWS_PER_BATCH = 102400)

The builder also owns the temp-table naming rule, which was duplicated with it:
`#name` lives in tempdb, so qualifying it as `[dbo].[#name]` sends the server
looking in the current database.

**Asserted on the text, because the text is invisible from SQL.** It goes to the
server and the server either accepts it or does not — every end-to-end test of a
load passes whether or not the hints are there, which is exactly how the missing
ROWS_PER_BATCH survived a release. test_insert_bulk_sql covers the four hint
combinations, both temp forms, and that 0 omits the hint rather than sending
`ROWS_PER_BATCH = 0`. It composes the type declarations from
GetSQLServerTypeDeclaration() rather than restating them: the subject here is
structure, not type rendering, and a test that restated it would fail on every
legitimate type change.

**Running `make test-cpp` surfaced three tests failing on main.**
test_ddl_translator, test_ctas_type_mapping and test_integer_codec all still
expected signed TINYINT to map to `tinyint`. Spec 057 deliberately widened it to
`smallint` — SQL Server's `tinyint` is its only unsigned integer (0..255) while
DuckDB's TINYINT is signed, so the narrow mapping silently turned -1 into 255 on
the wire. Expectations corrected, with the reason written next to them; all 13
standalone tests pass.

They went stale inside ONE spec, and the Makefile comment directly above them
says why this keeps happening: a test nobody runs is documentation that looks
like a guarantee. Spec 057 wired them into the Makefile and not into CI — and
could not, because no CI job builds the extension on a pull request at all
(ci.yml:161, issue #212), and these link against the built archive. D7 in the
spec now splits accordingly: the require-env gate is doable here, the standalone
gate is blocked behind #212, and the dependency is recorded rather than papered
over.

make test: 156 cases / 5009 assertions. make test-cpp: 13/13.

+++ test/cpp/codec/test_integer_codec.cpp
@@ -137,8 +137,13 @@ void TestFormatDdlTypeNameExpectedShapes() {
 	std::cout << "Test: FormatDdlTypeName returns SQL Server-canonical type names\n";
 
 	duckdb::mssql::CTASConfig cfg;
+	// Signed TINYINT widens to SMALLINT, deliberately (spec 057): SQL Server's
+	// `tinyint` is its ONLY unsigned integer (0..255) while DuckDB's TINYINT is
+	// signed (-128..127), so the narrow mapping silently turned -1 into 255 on
+	// the wire. UTINYINT is the type that actually matches `tinyint`, and the two
+	// lines below are what says the pair is not symmetric.
 	CHECK_EQ(duckdb::mssql::codec::integer::FormatDdlTypeName(LogicalType::TINYINT, cfg, DdlContext::CreateTable),
-			 std::string("TINYINT"));
+			 std::string("SMALLINT"));
 	CHECK_EQ(duckdb::mssql::codec::integer::FormatDdlTypeName(LogicalType::UTINYINT, cfg, DdlContext::CreateTable),
 			 std::string("TINYINT"));
 	CHECK_EQ(duckdb::mssql::codec::integer::FormatDdlTypeName(LogicalType::SMALLINT, cfg, DdlContext::CtasCreateTable),
